Исправление типа возвращаемого значения функции BmiService

+++ src/BmiService.java
@@ -1,5 +1,5 @@
 public class BmiService {
-    public double calculate(double weightKg, double heightMeters) {
+    public int calculate(double weightKg, double heightMeters) {
         double result = weightKg / (heightMeters * heightMeters);
         return (int) result;
